gitignore modified and app name passed

+++ backend/src/zango/apps/shared/tenancy/tasks.py
@@ -10,6 +10,8 @@ from django.conf import settings
 from django.core.management import call_command
 from django.utils import timezone
 
+import zango
+
 from .utils import DEFAULT_THEME_CONFIG, assign_policies_to_anonymous_user
 
 
@@ -44,13 +46,18 @@ def initialize_workspace(tenant_uuid, app_template_path=None):
             shutil.copytree(app_template_path, app_dir, dirs_exist_ok=True)
 
             shutil.rmtree(app_template_path)
-            shutil.rmtree(os.path.join(app_dir, "packages"))
+
+            if os.path.exists(os.path.join(app_dir, "packages")):
+                shutil.rmtree(os.path.join(app_dir, "packages"))
         else:
             # Creating app folder with the initial files
             template_directory = os.path.join(
                 os.path.dirname(__file__), "workspace_folder_template"
             )
-            cookiecutter_context = {"app_name": tenant.name}
+            cookiecutter_context = {
+                "app_name": tenant.name,
+                "zango_version": zango.__version__,
+            }
 
             cookiecutter.main.cookiecutter(
                 template_directory,
@@ -80,7 +87,9 @@ def initialize_workspace(tenant_uuid, app_template_path=None):
         else:
             return {"result": "failure", "error": "Failed, see error logs"}
     except Exception as e:
+        import traceback
+
         return {
             "result": "failure",
-            "error": str(e),
+            "error": traceback.format_exc(),
         }
